test: share journal::MockJournaler between librbd and rbd-mirror
(cherry picked from commit 04db7ddec661ef195077c3fab06f248705b17220)

Conflicts:
	src/test/librbd/test_mock_Journal.cc: MockJournaler moved to new file
